requirements.txt
flask==2.3.3
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0
PyJWT==2.8.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1